Simplify and refactor some project structures
