--- FlightTermination/tests/gcs_client/test_gcs_connection.py
@@ -1,9 +1,0 @@
-import unittest
-
-
-class TestGCSConnection(unittest.TestCase):
-    def test_connect_to_drone(self):
-        pass
-
-    def test_send_command_to_drone(self):
-        pass
