+++ flight_termination/flight_termination.py
@@ -1,0 +1,16 @@
+def begin_flight_termination():
+    pre_flight_termination()
+    # Set flight to Hold for some time.
+    # Then start flight termination.
+
+
+def pre_flight_termination():
+    pass
+
+
+def terminate_flight():
+    pass
+
+
+def post_flight_termination():
+    pass
