+++ flight_termination/run.sh
