+++ flight_termination/setup.sh
