+++ navigation/connection.py
@@ -203,24 +203,6 @@ def request_message_interval(
     )
 
 
-def begin_flight_termination():
-    pre_flight_termination()
-    # Set flight to Hold for some time.
-    # Then start flight termination.
-
-
-def pre_flight_termination():
-    pass
-
-
-def terminate_flight():
-    pass
-
-
-def post_flight_termination():
-    pass
-
-
 def connection_to_autopilot(conn_string, baudrate):
     return connection_to_mavlink_system(AUTOPILOT, conn_string, baudrate, 0, 0)
